Fix mismatched BeginDisabled/EndDisabled
On the frame where you click the stop button, the first `if` will evaluate to false (because a song is playing), but then the second `if` will evaluate to true (since a song is now no longer playing), meaning we call EndDisabled without having called BeginDisabled.

Switched to passing the predicate into the BeginDisabled function, similar to the play button.